+++ src/main/java/org/kumoricon/registration/BaseDataService.java
@@ -314,9 +314,7 @@ public class BaseDataService {
 
         // Create badge types with security restrictions below
         List<Badge> vipBadges = new ArrayList<>();
-        vipBadges.add(BadgeFactory.createBadge("VIP-Cumulus", BadgeType.VIP, "Weekend", "#000000", 250, 250, 250));
-        vipBadges.add(BadgeFactory.createBadge("VIP-Altocumulus", BadgeType.VIP, "Weekend", "#000000", 450, 450, 450));
-        vipBadges.add(BadgeFactory.createBadge("VIP-Cumulonimbus", BadgeType.VIP, "Weekend", "#000000", 600, 600, 600));
+        vipBadges.add(BadgeFactory.createBadge("VIP-Primary", BadgeType.VIP, "Weekend", "#000000", 500, 500, 500));
 
         for (Badge vip : vipBadges) {
             log.info("Creating badge {}", vip.getName());